Ajustes de navegacion y autorizacion

+++ flowboard_backend/src/controllers/ProjectController.ts
@@ -5,6 +5,7 @@ export class ProjectController {
 
     static createProject = async (req: Request, res: Response) => {
         const project = new Project(req.body)
+        project.manager = req.user.id
         try {
             await project.save()
             res.send('Project created successfully')
@@ -15,7 +16,11 @@ export class ProjectController {
 
     static getAllProjects = async (req: Request, res: Response) => {
         try {   
-            const projects = await Project.find({})
+            const projects = await Project.find({
+                $or: [
+                    { manager: {$in: req.user.id}}
+                ]
+            })
             res.json(projects)
         } catch (error) {
             res.status(500).json({ message: error.message })
@@ -27,6 +32,10 @@ export class ProjectController {
             const project = await Project.findById(req.params.id).populate('tasks')
             if (!project) {
                 res.status(404).json({ message: 'Project not found' })
+                return
+            }
+            if(project.manager.toString() !== req.user.id) {
+                res.status(401).json({message: 'Unauthorized'})
                 return
             }
             res.json(project)
@@ -44,6 +53,12 @@ export class ProjectController {
                 res.status(404).json({ message: 'Project not found' })
                 return
             }
+
+            if(project.manager.toString() !== req.user.id) {
+                res.status(401).json({message: 'Only the manager can update this project'})
+                return
+            }
+
             project.projectName = projectName
             project.clientName = clientName
             project.description = description
@@ -62,6 +77,12 @@ export class ProjectController {
                 res.status(404).json({ message: 'Project not found' })
                 return
             }
+
+            if(project.manager.toString() !== req.user.id) {
+                res.status(401).json({message: 'Only the manager can update this project'})
+                return
+            }
+            
             await project.deleteOne()
             res.send('Project deleted successfully')
         } catch (error) {
